+++ src/main/java/com/ecommerceapi/ecommerceapi/repository/ProductRepository.java
@@ -2,12 +2,13 @@ package com.ecommerceapi.ecommerceapi.repository;
 
 import com.ecommerceapi.ecommerceapi.entity.Product;
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.data.mongodb.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends MongoRepository<Product, Integer> {
+public interface ProductRepository extends MongoRepository<Product, Long> {
 
     Product findByProductName(String productName);
